Separated english and russian tests
