+++ tests/helpers.go
@@ -1,0 +1,13 @@
+package tests
+
+import "time"
+
+// test helper
+func smallSubTime(d time.Duration) string {
+	return time.Now().Add(d).Format("2006-01-02 15:04:05")
+}
+
+// test helper
+func bigSubTime(years int, months int, days int) string {
+	return time.Now().AddDate(-years, -months, -days).Format("2006-01-02 15:04:05")
+}
